Modify model/generateOrderNumber.py app.py model/registerOrder.py
