+++ app.py
@@ -39,16 +39,6 @@ app.state.jwt = my_jwt
 
 
 
-def generateOrderNumber(primeOutput:PrimeOutput):
-    today = datetime.datetime.now().strftime("%Y%m%d")
-    string = primeOutput.rec_trade_id
-    number = ""
-    for ch in string[:8:-1]:
-        number += str(ord(ch))
-    number = today + number
-    return number
-
-
 def orderBooking(userInfo: UserInfo,orderInput:OrderInput):
     # 取出資料
     userId = userInfo.data.id
