+++ model/registerOrder.py
@@ -1,5 +1,6 @@
 from fastapi.responses import JSONResponse
 from model.payByPrime import payByPrime
+from model.generateOrderNumber import generateOrderNumber
 from model.OrderModel import OrderInput, OrderOutput, OrderInfo, Payment
 from model.UserModel import UserInfo
 from model.ResponseModel import Error
